% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_words.R
\name{compare_words}
\alias{compare_words}
\title{Compare Words}
\usage{
compare_words(word_vector, sentiment_words)
}
\arguments{
\item{word_vector}{A vector of preprocessed words.}

\item{sentiment_words}{A vector of sentiment words (supports prefixes/suffixes).}
}
\value{
The count of matched words.
}
\description{
Matches words in a text vector to positive or negative word lists.
}
\examples{
compare_words(c("hope", "nauseous"), c("hope", "cooperat*"))
}
