% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_sentiment.R
\name{summarize_sentiment}
\alias{summarize_sentiment}
\title{Summarize Sentiment}
\usage{
summarize_sentiment(words, positive_words, negative_words)
}
\arguments{
\item{words}{A vector of preprocessed words.}

\item{positive_words}{Positive sentiment words.}

\item{negative_words}{Negative sentiment words.}
}
\value{
A list with counts and ratio.
}
\description{
Summarizes positive and negative word counts and calculates a sentiment ratio.
}
\examples{
summarize_sentiment(c("hope", "nauseous"), c("hope"), c("nauseous"))
}
